Refactor De-Flask
