+++ assets/JS/aos.js
@@ -1,0 +1,13 @@
+// // OR IF ONLY WANTING TO APPLY THE ANIMATION ONCE:
+const observer = new IntersectionObserver((entries) => {
+  entries.forEach((entry) => {
+    if (entry.isIntersecting) {
+      entry.target.classList.add("visible");
+      observer.unobserve(entry.target);
+    }
+  });
+});
+
+
+const hiddenElements = document.querySelectorAll(".hidden");
+hiddenElements.forEach((element) => observer.observe(element));
