+++ assets/css/styles.css
@@ -1,0 +1,228 @@
+html,
+body,
+*,
+:after,
+:before {
+  padding: 0px;
+  margin: 0px;
+  box-sizing: border-box;
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  line-height: 25px;
+  letter-spacing: 0.8px;
+}
+h1 {
+  font-family: Garamond, Baskerville, Baskerville Old Face, Hoefler Text,
+    Times New Roman, serif !important;
+  font-weight: 900;
+  font-size: clamp(5vw, 5vw, 200px);
+  color: #fff;
+}
+h2,
+.auto-type {
+  font-weight: 700;
+  line-height: 1.1;
+}
+h3 {
+  font-weight: 700;
+  line-height: 1.1;
+}
+p {
+  line-height: 25px;
+  letter-spacing: 0.8px;
+}
+
+/* Header */
+header {
+  display: grid;
+  position: fixed;
+  padding: 10px;
+  top: 0;
+  z-index: 2;
+  width: 100%;
+  text-align: center;
+  background-color: #000;
+  background-color: transparent;
+  transition: background-color 0.5s ease;
+  color: #fff;
+}
+
+header nav a {
+  text-decoration: none;
+  padding: 10px 10px;
+  color: inherit;
+  margin: 5px 10px;
+}
+header nav a:hover {
+  /* EY Yellow */
+  color: #ffe600;
+}
+
+/* Landing Page */
+
+#landing-page {
+  width: 100% !important;
+  position: relative;
+  align-items: center;
+  justify-content: center;
+}
+section .title-box {
+  position: absolute;
+  top: 40%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  text-align: center;
+  z-index: 1;
+}
+
+.landing-video {
+  width: 100%;
+  z-index: -1;
+  overflow: hidden;
+}
+/* Profile */
+.link-icons {
+  text-decoration: none;
+  color: #000;
+  margin: auto 5px;
+}
+.link-icons:hover {
+  color: darkgrey;
+}
+
+/* Page1 */
+.page {
+  max-width: 1100px;
+  margin: auto;
+}
+
+.page-title {
+  margin-top: 100px;
+  text-align: center;
+  font-size: 3rem;
+}
+.template-title-example {
+  font-size: 3rem;
+  margin: 50px auto;
+}
+.resume-row {
+  padding: 50px 20px;
+}
+.resume-row-label {
+  padding: 20px 0;
+}
+row {
+  margin-left: -15px;
+  margin-right: -15px;
+}
+.resume-photo {
+  vertical-align: middle;
+  text-align: center;
+  max-width: 70% !important;
+  margin: auto;
+}
+.template-example {
+  margin-left: 30px;
+  margin-bottom: 20px;
+}
+
+.skills-photo {
+  max-height: 120px;
+  width: auto;
+  margin: auto;
+}
+.skill-name {
+  margin-top: 20px;
+  font-size: 20px;
+  font-weight: 700;
+}
+
+footer {
+  margin: 30px auto 20px;
+}
+
+@media (max-width: 767px) {
+  .template-example {
+    margin-left: 0;
+  }
+  .resume-photo,
+  canvas {
+    margin-bottom: 50px;
+  }
+  .template-title-example {
+    margin: auto;
+    text-align: center;
+  }
+}
+
+
+#particles-js {
+  width: 100%;
+  height: 90vh;
+  background-color: #000;
+  background-image: url("");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: 50% 50%;
+}
+canvas {
+  display: block;
+}
+
+/* Animate on Scroll */
+.hidden {
+    opacity: 0;
+    filter: blur(5px);
+    transform: translateY(100%);
+    transition: 0;
+
+}
+@media (prefers-reduced-motion: no-preference) {
+  .hidden {
+    transition: all 0.6s;
+  }
+}
+
+.visible {
+    opacity: 1;
+    filter: blur(0);
+    transform: translateY(0);
+}
+
+.link-icons:nth-child(1) {
+  transition-delay: 100ms;
+}
+.link-icons:nth-child(2) {
+  transition-delay: 200ms;
+}
+.link-icons:nth-child(3) {
+  transition-delay: 300ms;
+}
+
+.logo:nth-child(1) {
+    transition-delay: 100ms;
+}
+.logo:nth-child(2) {
+    transition-delay: 150ms;
+}
+.logo:nth-child(3) {
+    transition-delay: 200ms;
+}
+.logo:nth-child(4) {
+    transition-delay: 250ms;
+}
+.logo:nth-child(5) {
+    transition-delay: 300ms;
+}
+.logo:nth-child(6) {
+    transition-delay: 350ms;
+}
+.logo:nth-child(7) {
+    transition-delay: 400ms;
+} 
+.logo:nth-child(8) {
+    transition-delay: 450ms;
+} 
+.logo:nth-child(9) {
+    transition-delay: 500ms;
+}
